+++ src/1.2-stock-span-lib.h
@@ -1,4 +1,4 @@
 #include <stddef.h>
 
-void stock_span_stack (int *quotes, size_t *spans, size_t length);
-void stock_span_simple (int *quotes, size_t *spans, size_t length);
+void stock_span_stack (double *quotes, size_t *spans, size_t length);
+void stock_span_simple (double *quotes, size_t *spans, size_t length);
